move common histogramming routines to client lib, DQMEDAnalyzers to the RPCMonitorDigi package
