+++ DQM/RPCMonitorClient/BuildFile.xml
@@ -1,20 +1,11 @@
-<flags EDM_PLUGIN="1"/>
-<use name="boost"/>
-<use name="rootcore"/>
-<use name="rootgraphics"/>
 <use name="DQMServices/Core"/>
-<use name="FWCore/Framework"/>
-<use name="FWCore/ParameterSet"/>
-<use name="FWCore/ServiceRegistry"/>
-<use name="FWCore/Utilities"/>
 <use name="Geometry/RPCGeometry"/>
-<use name="Geometry/CommonDetUnit"/>
-<use name="Geometry/Records"/>
-<use name="EventFilter/RPCRawToDigi"/>
-<use name="DataFormats/RPCDigi"/>
-<use name="DataFormats/FEDRawData"/>
-<use name="DataFormats/MuonDetId"/>
 <use name="CondFormats/RPCObjects"/>
 <use name="CondFormats/DataRecord"/>
 <use name="CondFormats/RunInfo"/>
 <use name="CondCore/PopCon"/>
+<use name="EventFilter/RPCRawToDigi"/>
+<export>
+  <lib name="1"/>
+</export>
+
