+++ DQM/RPCMonitorClient/plugins/BuildFile.xml
@@ -1,0 +1,6 @@
+<use name="DQM/RPCMonitorClient"/>
+<use name="DQMServices/Core"/>
+<use name="Geometry/RPCGeometry"/>
+<library name="DQMRPCMonitorClient_plugins" file="*.cc">
+  <flags EDM_PLUGIN="1"/>
+</library>
