+++ DQM/RPCMonitorClient/plugins/SealModule.cc
@@ -1,14 +1,4 @@
-#include "FWCore/PluginManager/interface/ModuleDef.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
-
-#include <DQM/RPCMonitorClient/interface/RPCFEDIntegrity.h>
-DEFINE_FWK_MODULE(RPCFEDIntegrity);
-
-#include <DQM/RPCMonitorClient/interface/RPCMonitorRaw.h>
-DEFINE_FWK_MODULE(RPCMonitorRaw);
-
-#include <DQM/RPCMonitorClient/interface/RPCMonitorLinkSynchro.h>
-DEFINE_FWK_MODULE(RPCMonitorLinkSynchro);
 
 //General Client
 #include <DQM/RPCMonitorClient/interface/RPCDqmClient.h>
@@ -16,9 +6,6 @@ DEFINE_FWK_MODULE(RPCDqmClient);
 
 #include <DQM/RPCMonitorClient/interface/RPCRecHitProbabilityClient.h>
 DEFINE_FWK_MODULE(RPCRecHitProbabilityClient);
-
-#include <DQM/RPCMonitorClient/interface/RPCChamberQuality.h>
-DEFINE_FWK_MODULE(RPCChamberQuality);
 
 #include <DQM/RPCMonitorClient/interface/RPCDcsInfoClient.h>
 DEFINE_FWK_MODULE(RPCDcsInfoClient);
