+++ DQM/RPCMonitorDigi/src/RPCMonitorRaw.cc
@@ -1,4 +1,4 @@
-#include "DQM/RPCMonitorClient/interface/RPCMonitorRaw.h"
+#include "DQM/RPCMonitorDigi/interface/RPCMonitorRaw.h"
 
 #include "DQMServices/Core/interface/DQMStore.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
@@ -11,7 +11,7 @@
 #include "CondFormats/RPCObjects/interface/LinkBoardElectronicIndex.h"
 
 #include "DataFormats/RPCDigi/interface/ReadoutError.h"
-#include "DQM/RPCMonitorClient/interface/RPCRawDataCountsHistoMaker.h"
+#include "DQM/RPCMonitorDigi/interface/RPCRawDataCountsHistoMaker.h"
 
 #include "TH1F.h"
 #include "TH2F.h"
